+++ build.gradle
@@ -1,18 +1,8 @@
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-    dependencies {
-        classpath("org.springframework.boot:spring-boot-gradle-plugin:2.1.7.RELEASE")
-    }
-}
-
 plugins {
     id 'java'
-    id 'war'
     id 'idea'
     id 'org.springframework.boot' version '2.1.7.RELEASE'
-    id "io.spring.dependency-management" version "1.0.8.RELEASE"
+    id 'io.spring.dependency-management' version '1.0.8.RELEASE'
 }
 
 group 'org.neorg'
@@ -22,12 +12,16 @@ repositories {
     mavenCentral()
 }
 
-sourceCompatibility = 11
+sourceCompatibility = JavaVersion.VERSION_11
+
+repositories {
+    mavenCentral()
+}
 
 dependencies {
     compile group: 'org.springframework.boot', name: 'spring-boot-starter-web', version: '2.1.7.RELEASE'
     compile group: 'io.springfox', name: 'springfox-swagger2', version: '2.9.2'
     compile group: 'io.springfox', name: 'springfox-swagger-ui', version: '2.9.2'
-    providedCompile group: 'org.projectlombok', name: 'lombok', version: '1.18.8'
+    compileOnly group: 'org.projectlombok', name: 'lombok', version: '1.18.8'
+    annotationProcessor 'org.projectlombok:lombok:1.18.8'
 }
-
