+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,5 +1,6 @@
+#Fri Sep 06 17:47:04 MSK 2019
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-5.6-bin.zip
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-5.6-all.zip
