--- src/main/java/testSystem/hello/HelloWorldController.java
@@ -1,17 +1,0 @@
-package testSystem.hello;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HelloWorldController {
-
-    @Autowired
-    HelloWorldService helloWorldService;
-
-    @GetMapping("/hello")
-    public String helloWorld() {
-        return helloWorldService.makeHelloString();
-    }
-}
