+++ src/main/java/testSystem/hello/controllers/HelloWorldController.java
@@ -1,0 +1,18 @@
+package testSystem.hello.controllers;
+
+import lombok.AllArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+import testSystem.hello.services.HelloWorldService;
+
+@RestController
+@AllArgsConstructor
+public class HelloWorldController {
+
+    private final HelloWorldService helloWorldService;
+
+    @GetMapping("/hello")
+    public String helloWorld() {
+        return helloWorldService.getHelloWorldMessage();
+    }
+}
