+++ src/main/java/testSystem/hello/services/HelloWorldService.java
@@ -1,0 +1,14 @@
+package testSystem.hello.services;
+
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@AllArgsConstructor
+public class HelloWorldService {
+    private static final String HELLO_WORLD = "Hello, World!";
+
+    public String getHelloWorldMessage() {
+        return HELLO_WORLD;
+    }
+}
